+++ src/main/java/com/nhnacademy/store99/bookstore/book/service/BookService.java
@@ -14,5 +14,9 @@ public interface BookService {
 
     BookResponse getBookDataById(Long bookId);
 
+    void deleteBook(Long bookId);
+
+    void restoreBook(Long bookId);
+
     void plusViewCnt(Long bookId);
 }
